Update parsec service to version 1.0.0
Updated the bitbake files to use version 1.0.0 of Parsec

Added recipes for the latest version of parsec-tool.

Update README.md to have a section about how to change to a newer version
of parsec.

Updated the initialisation phase to work with a read-only file system.  This
is needed for SoftHSM configurations where the config.toml file gets updated
during initialisation. This gets written to /userdata/parsec instead of being
updated in place in /etc/parsec.